requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
environ==1.0
include==0.2.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
uuid==1.30
whitenoise==6.4.0
